+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Remove `deprecated` attribute from `ISO`
+
 ## HCloud v1.7.2 (2023-10-16)
 
 - Add `deprecation` attribute to `ISO`